refactor ds
